Extract OAuth flows, scope resolution, and response filtering from api/credentials.py
Pure relocation with zero semantic change:
- api/credentials_oauth_flows.py: pending OAuth state issue/consume bound to
  the authenticated dashboard user (_PendingOAuthState).
- api/dashboard_credential_scope.py: execution-scope override parsing, agent
  scope resolution, dashboard execution identity, and credential-management
  authorization checks.
- api/credentials_target.py: RequestCredentialsTarget resolution and
  load/save/delete store routing for resolved targets.
- api/credentials_oauth_policy.py: OAuth credential service classification and
  dashboard read/write/copy rules, including client-config normalization.
- api/credential_responses.py: response security filtering.

api/credentials.py keeps route handlers plus the _DashboardCredentialAccess
delegation facade. Consumers (api/oauth.py, api/integrations.py,
api/homeassistant_integration.py, api/tools.py) import from the new modules;
no re-exports. Tests updated for import paths and mock targets only.
